Move GATT server into a FGS
