--- samples/connectivity/bluetooth/ble/src/main/java/com/example/platform/connectivity/bluetooth/ble/GATTServerSample.kt
@@ -1,345 +1,0 @@
-/*
- * Copyright 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.platform.connectivity.bluetooth.ble
-
-import android.Manifest
-import android.annotation.SuppressLint
-import android.app.Service
-import android.bluetooth.BluetoothDevice
-import android.bluetooth.BluetoothGatt
-import android.bluetooth.BluetoothGattCharacteristic
-import android.bluetooth.BluetoothGattServer
-import android.bluetooth.BluetoothGattServerCallback
-import android.bluetooth.BluetoothGattService
-import android.bluetooth.BluetoothManager
-import android.bluetooth.le.AdvertiseCallback
-import android.bluetooth.le.AdvertiseData
-import android.bluetooth.le.AdvertiseSettings
-import android.bluetooth.le.BluetoothLeAdvertiser
-import android.content.Context
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.content.pm.ServiceInfo
-import android.os.Build
-import android.os.IBinder
-import android.os.ParcelUuid
-import androidx.annotation.RequiresPermission
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.core.app.ActivityCompat
-import androidx.core.app.NotificationChannelCompat
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-import androidx.core.content.ContextCompat
-import androidx.core.content.getSystemService
-import com.google.android.catalog.framework.annotations.Sample
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
-import java.util.UUID
-
-
-@Sample(
-    name = "Create a GATT server",
-    description = "Shows how to create a GATT server and communicate with the GATT client",
-    documentation = "https://developer.android.com/reference/android/bluetooth/BluetoothGattServer",
-    tags = ["bluetooth"],
-)
-@Composable
-fun GATTServerSample() {
-    // In addition to the Bluetooth permissions we also need the BLUETOOTH_ADVERTISE from Android 12
-    val extraPermissions = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-        setOf(Manifest.permission.BLUETOOTH_ADVERTISE)
-    } else {
-        emptySet()
-    }
-    BluetoothSampleBox(extraPermissions = extraPermissions) { adapter ->
-        if (adapter.isMultipleAdvertisementSupported) {
-            GATTServerScreen()
-        } else {
-            Text(text = "Devices does not support multi-advertisement")
-        }
-    }
-}
-
-@Composable
-internal fun GATTServerScreen() {
-    val context = LocalContext.current
-    var enableServer by remember {
-        mutableStateOf(true)
-    }
-    val logs by GATTServerSampleService.serverLogsState.collectAsState()
-
-    LaunchedEffect(enableServer) {
-        val intent = Intent(context, GATTServerSampleService::class.java)
-        if (enableServer) {
-            ContextCompat.startForegroundService(context, intent)
-        } else {
-            context.stopService(intent)
-        }
-    }
-
-    Column(
-        Modifier
-            .fillMaxSize()
-            .verticalScroll(rememberScrollState())
-            .padding(16.dp),
-    ) {
-        Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-            Button(onClick = { enableServer = !enableServer }) {
-                Text(text = if (enableServer) "Stop Server" else "Start Server")
-            }
-
-            Button(onClick = { enableServer = !enableServer }) {
-                Text(text = if (enableServer) "Stop Server" else "Start Server")
-            }
-        }
-        Text(text = logs)
-    }
-}
-
-// Random UUID for our service known between the client and server to allow communication
-val SERVICE_UUID: UUID = UUID.fromString("00002222-0000-1000-8000-00805f9b34fb")
-
-// Same as the service but for the characteristic
-val CHARACTERISTIC_UUID: UUID = UUID.fromString("00001111-0000-1000-8000-00805f9b34fb")
-
-internal class GATTServerSampleService : Service() {
-
-    companion object {
-        val serverLogsState: MutableStateFlow<String> = MutableStateFlow("")
-        val isServerRunning = MutableStateFlow(false)
-        val advertiseEnabled = MutableStateFlow(true)
-
-        private const val CHANNEL = "gatt_server_channel"
-
-        private fun GATTServerSampleService.startForeground() {
-            createNotificationChannel()
-
-            val notification = NotificationCompat.Builder(this, CHANNEL)
-                .setSmallIcon(applicationInfo.icon)
-                .setContentTitle("GATT Server")
-                .setContentText("Running...")
-                .build()
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                startForeground(
-                    100,
-                    notification,
-                    ServiceInfo.FOREGROUND_SERVICE_TYPE_CONNECTED_DEVICE,
-                )
-            } else {
-                startForeground(100, notification)
-            }
-        }
-
-        private fun Context.createNotificationChannel() {
-            val channel =
-                NotificationChannelCompat.Builder(
-                    CHANNEL,
-                    NotificationManagerCompat.IMPORTANCE_HIGH,
-                )
-                    .setName("GATT Server channel")
-                    .setDescription("Channel for the GATT server sample")
-                    .build()
-            NotificationManagerCompat.from(this).createNotificationChannel(channel)
-        }
-    }
-
-    private val manager: BluetoothManager by lazy {
-        applicationContext.getSystemService()!!
-    }
-    private val advertiser: BluetoothLeAdvertiser
-        get() = manager.adapter.bluetoothLeAdvertiser
-
-    private val service =
-        BluetoothGattService(SERVICE_UUID, BluetoothGattService.SERVICE_TYPE_PRIMARY).also {
-            it.addCharacteristic(
-                BluetoothGattCharacteristic(
-                    CHARACTERISTIC_UUID,
-                    BluetoothGattCharacteristic.PROPERTY_WRITE or BluetoothGattCharacteristic.PROPERTY_READ,
-                    BluetoothGattCharacteristic.PERMISSION_WRITE or BluetoothGattCharacteristic.PERMISSION_READ,
-                ),
-            )
-        }
-
-    private lateinit var server: BluetoothGattServer
-
-    private val scope = CoroutineScope(SupervisorJob())
-
-    override fun onCreate() {
-        super.onCreate()
-        serverLogsState.value = ""
-        if (ActivityCompat.checkSelfPermission(this, Manifest.permission.BLUETOOTH_CONNECT)
-            == PackageManager.PERMISSION_GRANTED
-        ) {
-            startForeground()
-            serverLogsState.value += "Opening GATT server\n"
-            server = manager.openGattServer(applicationContext, SampleServerCallback())
-            server.addService(service)
-            isServerRunning.value = true
-
-            if (ActivityCompat.checkSelfPermission(this, Manifest.permission.BLUETOOTH_ADVERTISE)
-                == PackageManager.PERMISSION_GRANTED
-            ) {
-                scope.launch {
-                    advertiseEnabled.collect {
-                        if (it) {
-                            serverLogsState.value += "Enable advertising\n"
-                            startAdvertising()
-                        } else {
-                            serverLogsState.value += "Disable advertising\n"
-                            advertiser.stopAdvertising(SampleAdvertiseCallback)
-                        }
-                    }
-                }
-            } else {
-                serverLogsState.value += "Missing advertise permission\n"
-            }
-        } else {
-            serverLogsState.value += "Missing connect permission\n"
-            stopSelf()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        isServerRunning.value = false
-        if (ActivityCompat.checkSelfPermission(this, Manifest.permission.BLUETOOTH_ADVERTISE)
-            == PackageManager.PERMISSION_GRANTED
-        ) {
-            advertiser.stopAdvertising(SampleAdvertiseCallback)
-        }
-        server.close()
-        scope.cancel()
-        serverLogsState.value += "Server destroyed\n"
-    }
-
-    @SuppressLint("InlinedApi")
-    @RequiresPermission(Manifest.permission.BLUETOOTH_ADVERTISE)
-    private fun startAdvertising() {
-        val settings = AdvertiseSettings.Builder()
-            .setAdvertiseMode(AdvertiseSettings.ADVERTISE_MODE_BALANCED)
-            .setTxPowerLevel(AdvertiseSettings.ADVERTISE_TX_POWER_MEDIUM)
-            .setConnectable(true)
-            .setTimeout(0)
-            .build()
-
-        val data = AdvertiseData.Builder()
-            .setIncludeDeviceName(true)
-            .addServiceUuid(ParcelUuid(SERVICE_UUID))
-            .build()
-
-        advertiser.startAdvertising(settings, data, SampleAdvertiseCallback)
-    }
-
-    inner class SampleServerCallback : BluetoothGattServerCallback() {
-
-        @SuppressLint("MissingPermission")
-        override fun onConnectionStateChange(
-            device: BluetoothDevice,
-            status: Int,
-            newState: Int,
-        ) {
-            serverLogsState.value += "\nConnection state change: ${newState.toConnectionStateString()}." +
-                    " New device: ${device.name} ${device.address}"
-            // You should keep a list of connected device to manage them
-        }
-
-        @SuppressLint("MissingPermission")
-        override fun onCharacteristicWriteRequest(
-            device: BluetoothDevice,
-            requestId: Int,
-            characteristic: BluetoothGattCharacteristic,
-            preparedWrite: Boolean,
-            responseNeeded: Boolean,
-            offset: Int,
-            value: ByteArray,
-        ) {
-            serverLogsState.value += "\nCharacteristic Write request: $requestId\n" +
-                    "Data: ${String(value)} (offset $offset)"
-            // Here you should apply the write of the characteristic and notify connected
-            // devices that it changed
-
-            // If response is needed reply to the device that the write was successful
-            if (responseNeeded) {
-                server.sendResponse(
-                    device,
-                    requestId,
-                    BluetoothGatt.GATT_SUCCESS,
-                    0,
-                    null,
-                )
-            }
-        }
-
-        @SuppressLint("MissingPermission")
-        override fun onCharacteristicReadRequest(
-            device: BluetoothDevice?,
-            requestId: Int,
-            offset: Int,
-            characteristic: BluetoothGattCharacteristic?,
-        ) {
-            super.onCharacteristicReadRequest(device, requestId, offset, characteristic)
-            serverLogsState.value += "\nCharacteristic Read request: $requestId (offset $offset)"
-            val data = serverLogsState.value.toByteArray()
-            val response = data.copyOfRange(offset, data.size)
-            server.sendResponse(
-                device,
-                requestId,
-                BluetoothGatt.GATT_SUCCESS,
-                offset,
-                response,
-            )
-        }
-
-        override fun onMtuChanged(device: BluetoothDevice?, mtu: Int) {
-            serverLogsState.value += "\nMTU change request: $mtu"
-        }
-    }
-
-    object SampleAdvertiseCallback : AdvertiseCallback() {
-        override fun onStartSuccess(settingsInEffect: AdvertiseSettings?) {
-            serverLogsState.value += "\nStarted advertising"
-        }
-
-        override fun onStartFailure(errorCode: Int) {
-            serverLogsState.value += "\nFailed to start advertising: $errorCode"
-        }
-    }
-
-    override fun onBind(intent: Intent?): IBinder? = null
-}
